[JBIDE-15744]: Added support for List<IEnvironmentVariable> to IApplication interface, as well as other side-effect changes.
The source changes are associated with JBIDE-15744.
The public interface method changes for IApplication are listed below:
A) public List getEnvironmentVariables(); This is change listing 3, in the change log, below.
B) public List getEnvironmentVaraiblesList()
C) public Map getEnvironmentVariableMap()

Change 3 made sense because we don't actually do any work on the list that is returned, and
really just require a glob of information.

Further support for this change will need to be made within jbosstools-openshift to reference
the references List which are required.
This work should be submitted with the changes in JBIDE-15766 as seen here: https://issues.jboss.org/browse/JBIDE-15766

Please reference the change log given below:

 1)Added getEnvironmentVariablesList(), and getEnvironmentVariablesMap()
    to IApplication interface and ApplicationResource.

 2)Refactored references to getEnvironmentVariables():Map<String,IEnvironmentVariable> to
    getEnvironmentVariablesMap().

 3)Refactored environmentVariableByName to environmentVariablesMap.

 4)Refactored getEnvironmentVariables():Map<String,IEnvironmentVariable>
    to return List<IEnvironmentVariable>.

 5)Refactored reference to getEnvironmentVariables(), to reference
    getEnvironmentVariablesMap() in ApplicationTestUtils

 6)Refactored reference to getEnvironmentVariables(), to reference
    getEnvironmentVariablesMap() in ApplicationResourceTest

 7)Refactored reference to getEnvironmentVariables(), to reference
    getEnvironmentVariablesMap() in ApplicationResourceIntegrationTest

 8)Refactored ApplicationResource.environmentVariablesByName to
	ApplicationResource.environmentVariablesMap

 9)Refactored Tests that reference "List" to use List<IEnvironmentVariable>.

 10)Updated comments in IApplication, ApplicationResource, and ApplicationResourceIntegrationTest.